working version bluetooth
This version works to establish bluetooth conection.
This has only been tested by using a bluetooth module and a led as output
There is still a bug: When you say something it will keep sending that and the string gets longer and longer the more you say.

so instead of:
Left
Right
Forward.

it does:
Left
Left Right
Left Right Forward